backend/requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.30.0
python-multipart==0.0.9
anthropic==0.40.0
librosa==0.10.2
numpy==1.26.4
crepe==0.0.16
pretty-midi==0.2.10
pydantic==2.9.0
python-dotenv==1.0.1
onnxruntime==1.19.0
scipy==1.13.0
soundfile==0.12.1
aiofiles==24.1.0
